split in to smaller files
