+++ src/backend/app/core/repository/__init__.py
@@ -15,7 +15,7 @@ from .document_repo import DocumentRepo
 from .code_elements.code_element_repo import CodeElementRepo
 from .code_elements.test_repo import TestRepo
 from .code_elements.play_ground_repo import PlayGroundRepo
-from .conversation_repo import ConversationRepo
+from .conversation import ConversationRepo
 
 
 class Repositories:
