+++ src/backend/app/core/repository/conversation/__init__.py
@@ -1,0 +1,3 @@
+from .repo import ConversationRepo
+
+__all__ = ["ConversationRepo"]
